chore(editor-state): split menus into multiple files
