+++ packages/editor/packages/editor-state/src/features/menu/menus/colorSchemeMenu.ts
@@ -1,0 +1,10 @@
+import type { MenuGenerator } from '../../../types';
+
+export const colorSchemeMenu: MenuGenerator = state => {
+	return state.colorSchemes.map(key => ({
+		title: key,
+		selector: 'editorSettings.colorScheme',
+		value: key,
+		close: false,
+	}));
+};
