+++ packages/editor/packages/editor-state/src/features/menu/menus/midiInfoMenu.ts
@@ -1,0 +1,14 @@
+import type { MenuGenerator } from '../../../types';
+
+export const midiInfoMenu: MenuGenerator = state => [
+	{ title: 'Inputs:', disabled: true, isSectionTitle: true },
+	...state.midi.inputs.map(input => ({
+		title: `${input.name || input.id}${input.manufacturer ? ` ${input.manufacturer}` : ``}`,
+		disabled: true,
+	})),
+	{ title: 'Outputs:', disabled: true, isSectionTitle: true },
+	...state.midi.outputs.map(output => ({
+		title: `${output.name || output.id}${output.manufacturer ? ` ${output.manufacturer}` : ``}`,
+		disabled: true,
+	})),
+];
